fix(v0.8.2): address per-lane review findings
- D-surfaces (CRITICAL): wrap tryWindowReport in try/catch so a memory.db
  error degrades to the legacy single-day path instead of crashing glean today.
- D-surfaces: add the windowRuns.length===0 coverage test (real DB, runs predate
  window) — the prior 0-burst test only hit the no-db guard.
- D-core: config max_unproductive + anti_spill_margin_minutes now .int() (a
  fractional threshold would never trip the integer counter) + reject-fractional tests.
- D-core: document that lock-busy/not-eligible in TRANSIENT_REASONS are defensive
  (only discovery-failed reaches the fold today).
- D-signal: correct the BLOCK-CAPTURE comment (per-task capture, not once-global).

398 passed + 2 skipped.

+++ src/lib/config.ts
@@ -15,10 +15,12 @@ const Schema = z.object({
     duration_hours: z.number().optional(),
     // v0.8.2 item 1: configurable circuit-breaker threshold (was the hard-coded
     // MAX_UNPRODUCTIVE = 3). Optional — defaults to 3 in runDrain when unset.
-    max_unproductive: z.number().optional(),
-    // v0.8.2 item 3: anti-spill pre-emptive margin in minutes. Optional —
-    // defaults to 15 in runDrain when unset.
-    anti_spill_margin_minutes: z.number().optional(),
+    // Integer: it's compared against the whole-number unproductive_reentries
+    // counter, so a fractional value (e.g. 3.7) would never trip the guard.
+    max_unproductive: z.number().int().optional(),
+    // v0.8.2 item 3: anti-spill pre-emptive margin in minutes (whole minutes).
+    // Optional — defaults to 15 in runDrain when unset.
+    anti_spill_margin_minutes: z.number().int().optional(),
   }).optional(),
 });
